// src/packages/types/carservice/Ser_Inv_StockIn.ts
export interface Ser_Inv_StockIn_Search {
  DealerCode: string;
  StockInNo: string; // Số phiếu nhập
  SupplierID: string; // NCC = 2951(HTC)
  TruckNo: string;
  StockDateFromTo?: Date[]; // Ngày nhập kho
  StockDateFrom?: string;
  StockDateTo?: string;
  OrderNo: string | boolean; // Số đơn hàng
  IsPending: string | boolean; // Mới tạo
  IsExecuting: string | boolean; // Tiến hành
  IsFinished: string | boolean; // Kết thúc
  IsAdjustment: string | boolean; // Điều chỉnh
  IsReject: string | boolean; // Hủy
  Ft_PageIndex: number;
  Ft_PageSize: number;
  FlagDataWH: boolean; // Lấy dữ liệu lịch sử
}